Rename "session_page" folder to "deck_page"
