+++ lib/src/pages/deck_page/error_card.dart
