+++ lib/src/pages/deck_page/result_section.dart
@@ -3,9 +3,9 @@ import 'package:ankigpt/src/models/anki_card.dart';
 import 'package:ankigpt/src/models/card_feedback.dart';
 import 'package:ankigpt/src/models/card_id.dart';
 import 'package:ankigpt/src/models/language.dart';
-import 'package:ankigpt/src/pages/session_page/error_card.dart';
-import 'package:ankigpt/src/pages/session_page/search_bar.dart';
-import 'package:ankigpt/src/pages/session_page/warning_card.dart';
+import 'package:ankigpt/src/pages/deck_page/error_card.dart';
+import 'package:ankigpt/src/pages/deck_page/search_bar.dart';
+import 'package:ankigpt/src/pages/deck_page/warning_card.dart';
 import 'package:ankigpt/src/pages/widgets/ankigpt_card.dart';
 import 'package:ankigpt/src/pages/widgets/card_feedback_dialog.dart';
 import 'package:ankigpt/src/pages/widgets/elevated_button.dart';
