+++ lib/src/pages/deck_page/warning_card.dart
